@pelletier 2023-05-19
I'm going to merge this as it's only in `unstable`, and I suspect may be lightly conflicting with @ImSingee's [work on `toml.Unmarshaler`](https://github.com/pelletier/go-toml/pull/870).