@hhildebrandt 2023-05-23
Interesting extension! Can you give an outlook when the comment support will be officially tagged? 
BTW - The current version is not buildable, because in go.mod go version 1.16 is pulled, but here a function from 1.17 is used: https://github.com/pelletier/go-toml/blob/44c1513ccd8d946f2be4e1fafd5675be00de0d6f/unmarshaler.go#LL1030C7-L1030C24