@pelletier 2023-05-23
@hhildebrandt released :) https://github.com/pelletier/go-toml/releases/tag/v2.0.8 and thank you for the heads up about PointerTo.